Common, 服务层移植

+++ HLL.HLX.BE.Application/Mobility/Videos/Dto/DeleteVideoInput.cs
@@ -11,6 +11,6 @@ namespace HLL.HLX.BE.Application.Mobility.Videos.Dto
     public class DeleteVideoInput : BaseInput
     {
         [Required]
-        public long? Id { get; set; }
+        public int? Id { get; set; }
     }
 }
