+++ HLL.HLX.BE.Core.Model/Tax/TaxCategory.cs
@@ -7,7 +7,7 @@ namespace HLL.HLX.BE.Core.Model.Tax
     /// <summary>
     /// Represents a tax category
     /// </summary>
-    public partial class TaxCategory : FullAuditedEntity<long, User>
+    public partial class TaxCategory : FullAuditedEntity<int, User>
     {
         /// <summary>
         /// Gets or sets the name
