+++ HLL.HLX.BE.Core.Model/Videos/IVideoRepository.cs
@@ -2,7 +2,7 @@
 
 namespace HLL.HLX.BE.Core.Model.Videos
 {
-    public interface IVideoRepository : IRepository<Video, long>
+    public interface IVideoRepository : IRepository<Video, int>
     {
     }
 }
